refactoring + new edit course page
