+++ src/test/java/com/spartaglobal/finalweek/pages/coursePages/EditCoursesPage.java
@@ -1,18 +1,23 @@
 package com.spartaglobal.finalweek.pages.coursePages;
 
 import com.spartaglobal.finalweek.interfaces.URLable;
-import com.spartaglobal.finalweek.pages.NavTemplate;
-import org.openqa.selenium.support.PageFactory;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
 
-public class EditCoursesPage extends NavTemplate implements URLable {
+import static com.spartaglobal.finalweek.base.TestBase.webDriver;
 
-    public EditCoursesPage() {
-        PageFactory.initElements(webDriver, this);
+public class EditCoursesPage extends ModifyCoursePage implements URLable {
+
+    private @FindBy(xpath = "//*[@id=\"edit-course-form\"]/div[2]/span[1]/button")
+    WebElement numberOfTrainersIncrement;
+    private @FindBy (xpath = "//*[@id=\"edit-course-form\"]/div[2]/span[2]/button/span")
+    WebElement numberOfTrainersDecrement;
+
+    public void incrementNumberOfTrainers(){
+        numberOfTrainersIncrement.click();
     }
 
-    @Override
-    public String getURL() {
-        return webDriver.getCurrentUrl();
+    public void decrementNumberOfTrainers(){
+        numberOfTrainersDecrement.click();
     }
-
 }
